+++ swagger-request-validator-core/src/test/java/com/atlassian/oai/validator/schema/SchemaValidatorTest.java
@@ -257,7 +257,17 @@ public class SchemaValidatorTest {
         final String value = "{\"name\": \"Moggy\", \"petType\": \"Cat\", \"huntingSkill\":\"clueless\"}";
 
         assertPass(classUnderTest.validate(value, schema));
+    }
 
+    @Test
+    public void validate_withDiscriminator_shouldPass_everyTime_whenInvokedMultipleTimes() {
+
+        final SchemaValidator classUnderTest = validatorWithAdditionalPropertiesIgnored("/oai/api-discriminator.yaml");
+        final Model schema = new RefModel("#/definitions/Pet");
+        final String value = "{\"name\": \"Moggy\", \"petType\": \"Cat\", \"huntingSkill\":\"clueless\"}";
+
+        assertPass(classUnderTest.validate(value, schema));
+        assertPass(classUnderTest.validate(value, schema));
     }
 
     @Test
@@ -268,7 +278,21 @@ public class SchemaValidatorTest {
         final String value = "{\"name\": \"Moggy\", \"petType\": \"Cat\", \"huntingSkill\":\"ruthless\"}";
 
         assertFail(classUnderTest.validate(value, schema), "validation.schema.discriminator");
+    }
 
+    @Test
+    public void validate_withDiscriminator_shouldFail_everyTime_whenInvokedMultipleTimes() {
+
+        final SchemaValidator classUnderTest = validatorWithAdditionalPropertiesIgnored("/oai/api-discriminator.yaml");
+        final Model schema = new RefModel("#/definitions/Pet");
+        final String value = "{\"name\": \"Moggy\", \"petType\": \"Cat\", \"huntingSkill\":\"ruthless\"}";
+
+        assertFail(classUnderTest.validate(value, schema), "validation.schema.discriminator");
+        assertFail(classUnderTest.validate(value, schema), "validation.schema.discriminator");
+        assertFail(classUnderTest.validate(value, schema), "validation.schema.discriminator");
+        assertFail(classUnderTest.validate(value, schema), "validation.schema.discriminator");
+        assertFail(classUnderTest.validate(value, schema), "validation.schema.discriminator");
+        assertFail(classUnderTest.validate(value, schema), "validation.schema.discriminator");
     }
 
     private SchemaValidator validatorWithAdditionalPropertiesIgnored(final String api) {